# swagger.json
{
    "swagger": "2.0",
    "info": {
        "version": "1.0.0",
        "title": "Saina Gallery RESTful API",
        "description": "SainaGallery API documentation"
    },
...
    "schemes": [
        "http"
    ],
    "paths": {
...
        "/api/province": {
            "get": {
                "tags": [
                    "api"
                ],
                "summary": "Get one Province",
                "description": "Get one Province",
                "operationId": "getoneprovinces",
                "parameters": [
                    {
                        "in": "query",
                        "type": "string",
                        "name": "id",
                        "description": "Province id"
                    }
                ],
                "responses": {
                    "405": {
                        "description": "Invalid input"
                    }
                },
                "security": [
                    {
                        "Authorization": []
                    }
                ]
            }
        },
...
        "/api/city": {
            "get": {
                "tags": [
                    "api"
                ],
                "summary": "Get one City",
                "description": "Get one City",
                "operationId": "getonecity",
                "parameters": [
                    {
                        "in": "query",
                        "type": "string",
                        "name": "id",
                        "description": "City id"
                    }
                ],
                "responses": {
                    "405": {
                        "description": "Invalid input"
                    }
                },
                "security": [
                    {
                        "Authorization": []
                    }
                ]
            }
        },
...
        "/api/gallery": {
            "get": {
                "tags": [
                    "api"
                ],
                "summary": "Get one Gallery",
                "description": "Get one Gallery",
                "operationId": "getonegallery",
                "parameters": [
                    {
                        "in": "query",
                        "type": "string",
                        "name": "id",
                        "description": "Gallery id"
                    }
                ],
                "responses": {
                    "405": {
                        "description": "Invalid input"
                    }
                },
                "security": [
                    {
                        "Authorization": []
                    }
                ]
            }
        },
...
        "/api/user": {
            "get": {
                "tags": [
                    "api"
                ],
                "summary": "Get One User",
                "description": "Get One User",
                "operationId": "getoneuser",
                "parameters": [
                    {
                        "in": "query",
                        "type": "string",
                        "name": "id"
                    }
                ],
                "responses": {
                    "405": {
                        "description": "Invalid input"
                    }
                },
                "security": [
                    {
                        "Authorization": []
                    }
                ]
            }
        },
...
    },
    "securityDefinitions": {
        "Authorization": {
            "type": "apiKey",
            "in": "header",
            "name": "Authorization"
        }
    },
...
}